requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
black==23.3.0
    # via -r requirements.in
blinker==1.6.2
    # via flask
click==8.1.3
    # via
    #   black
    #   flask
flake8==6.0.0
    # via -r requirements.in
flask==2.3.2
    # via -r requirements.in
isort==5.12.0
    # via -r requirements.in
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
packaging==23.1
    # via black
pathspec==0.11.1
    # via black
platformdirs==3.8.0
    # via black
pyactiveresource==2.2.2
    # via shopifyapi
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pyjwt==2.7.0
    # via shopifyapi
python-dotenv==1.0.0
    # via -r requirements.in
pyyaml==6.0
    # via shopifyapi
shopifyapi==12.3.0
    # via -r requirements.in
six==1.16.0
    # via
    #   pyactiveresource
    #   shopifyapi
werkzeug==2.3.6
    # via flask
